chore: deprecated firebaseDataHandler
